bt-664: * cherry-pick logout button
