+++ mobile/src/bundles/employer/screens/employer-onboarding/employer-onboarding.tsx
@@ -1,16 +1,14 @@
 import React from 'react';
 
-import { logout } from '~/bundles/auth/store/actions';
 import {
-    CommunityIcon,
-    Pressable,
+    LogoutButton,
     ScrollView,
     StatusBar,
     Text,
     View,
 } from '~/bundles/common/components/components';
-import { Color, IconName, TextCategory } from '~/bundles/common/enums/enums';
-import { useAppDispatch, useAppSelector } from '~/bundles/common/hooks/hooks';
+import { Color, TextCategory } from '~/bundles/common/enums/enums';
+import { useAppSelector } from '~/bundles/common/hooks/hooks';
 import { globalStyles } from '~/bundles/common/styles/styles';
 import { EmployerOnboardingForm } from '~/bundles/employer/components/components';
 import { useEmployerFormSubmit } from '~/bundles/employer/hooks/hooks';
@@ -19,7 +17,6 @@ import { type EmployerOnboardingFormDto } from '~/bundles/employer/types/types';
 import { styles } from './styles';
 
 const EmployerOnboarding: React.FC = () => {
-    const dispatch = useAppDispatch();
     const { onboardingData } = useAppSelector(({ common }) => common);
 
     const employerOnboardingData: EmployerOnboardingFormDto | null =
@@ -45,10 +42,6 @@ const EmployerOnboarding: React.FC = () => {
         void handleSubmit(payload);
     };
 
-    const handleLogout = (): void => {
-        void dispatch(logout());
-    };
-
     return (
         <>
             <StatusBar
@@ -72,13 +65,7 @@ const EmployerOnboarding: React.FC = () => {
                     <Text category={TextCategory.H4} style={[styles.title]}>
                         Create a profile to find a perfect match to your company
                     </Text>
-                    <Pressable onPress={handleLogout}>
-                        <CommunityIcon
-                            name={IconName.LOGOUT}
-                            size={30}
-                            color={Color.TEXT2}
-                        />
-                    </Pressable>
+                    <LogoutButton />
                 </View>
                 <Text category={TextCategory.H6} style={globalStyles.mb10}>
                     Please, fill out all the fields below, so we could verify
