+++ mobile/src/bundles/employer/screens/employer-profile/employer-profile.tsx
@@ -1,17 +1,15 @@
 import React from 'react';
 
-import { logout } from '~/bundles/auth/store/actions';
 import {
-    CommunityIcon,
-    Pressable,
+    LogoutButton,
     ScrollView,
     StatusBar,
     Text,
     VerificationMessage,
     View,
 } from '~/bundles/common/components/components';
-import { Color, IconName, TextCategory } from '~/bundles/common/enums/enums';
-import { useAppDispatch, useAppSelector } from '~/bundles/common/hooks/hooks';
+import { Color, TextCategory } from '~/bundles/common/enums/enums';
+import { useAppSelector } from '~/bundles/common/hooks/hooks';
 import { globalStyles } from '~/bundles/common/styles/styles';
 import { EmployerOnboardingForm } from '~/bundles/employer/components/components';
 import { useEmployerFormSubmit } from '~/bundles/employer/hooks/hooks';
@@ -20,7 +18,6 @@ import { type EmployerOnboardingFormDto } from '~/bundles/employer/types/types';
 import { styles } from './styles';
 
 const EmployerProfile: React.FC = () => {
-    const dispatch = useAppDispatch();
     const { onboardingData } = useAppSelector(({ common }) => common);
 
     const { isApproved } = onboardingData ?? {};
@@ -48,10 +45,6 @@ const EmployerProfile: React.FC = () => {
         void handleSubmit(payload);
     };
 
-    const handleLogout = (): void => {
-        void dispatch(logout());
-    };
-
     return (
         <>
             <StatusBar
@@ -77,13 +70,7 @@ const EmployerProfile: React.FC = () => {
                     ]}
                 >
                     {!isApproved && <VerificationMessage />}
-                    <Pressable onPress={handleLogout}>
-                        <CommunityIcon
-                            name={IconName.LOGOUT}
-                            size={30}
-                            color={Color.TEXT2}
-                        />
-                    </Pressable>
+                    <LogoutButton />
                 </View>
             </View>
